requirements.txt
python-telegram-bot==22.6
yt-dlp==2026.3.3
requests
